Build a offline solution that builds a RAG using your project files and retrieves the chunks before asking local LLM for response.
Updating Extensions with better handling  depending on project you are working on

Adding a chunker factory, a separate file that deals with providing the chucker depending on the file extension

Using the chunker factory to get better formatted chunks to be stored in vectorstore

Adding time, more logs for the processing of each file to understand the bottlenecks

Generating and comparing the hash files for the current directory to avoid re-constructing the vector store from scratch

Optimizing the chunking and creating a fingerprint for each chunk for better indexing and storing the metadata, which helps for ranking and context aware prompts and responses.

Updating Chunker factory weaponizing with detecting type, summarizing, formatting and extracting chunks involving updated regex for each language chunker.

V1- splitting the chunking logic, building rag into separate file, build_rag.py, keeping the streamlit ui purely in app.py